Mark Open311 contacts non_public according to service keywords
This allows the Open311 endpoint to include ‘private’ in the ‘keywords’ field
of a service definition in order for it to be marked as non_public on FMS.

NB existing categories may be updated to become non_public, but non_public
categories will never have their non_public flag unset by PopulateServiceList.
This is so any existing Open311 categories on FMS don’t suddenly become public.